surfer/requirements.txt
openai
python-dotenv
matplotlib
ollama
langchain_community